Rename RatekeeperData.actor.cpp to Ratekeeper.actor.cpp
